Refactor Commerce Class

+++ TallerJava/src/main/java/org/tallerJava/commerceModule/application/CommerceService.java
@@ -9,15 +9,15 @@ import java.util.List;
 public interface CommerceService {
     public boolean create(Commerce commerce);
     public boolean update(Commerce commerce);
-    public boolean updatePassword(int rut, String newPass);
-    public boolean delete(int rut);
+    public boolean updatePassword(long rut, String newPass);
+    public boolean delete(long rut);
 
-    public boolean createComplaint(int rut_commerce, String message);
-    public boolean createPos(int rut_commerce, Pos pos);
-    public boolean changePosStatus(int rut_commerce, Pos pos, boolean status);
+    public boolean createComplaint(long rut_commerce, String message);
+    public boolean createPos(long rut_commerce, Pos pos);
+    public boolean changePosStatus(long rut_commerce, Pos pos, boolean status);
 
 
-    public Commerce getByRut(int rut);
+    public Commerce getByRut(long rut);
     public List<Commerce> getAll();
 
 }
